Refactor the queue work and setup app
