+++ composer-laravel7.lock
@@ -4,7 +4,7 @@
         "Read more about it at https://getcomposer.org/doc/01-basic-usage.md#installing-dependencies",
         "This file is @generated automatically"
     ],
-    "content-hash": "44b5baa622d66677cfa71321c63be3fe",
+    "content-hash": "7ea23470da03428bbfa8947fc45dc735",
     "packages": [
         {
             "name": "carbonphp/carbon-doctrine-types",
@@ -5080,5 +5080,8 @@
         "php": "^7.2.5|~8.0.0"
     },
     "platform-dev": [],
+    "platform-overrides": {
+        "php": "7.2.10"
+    },
     "plugin-api-version": "2.3.0"
 }
