+++ src/App/Http/Controllers/Admin/Main_Controller.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Enpii_Base\App\Http\Controllers\Admin;
 
+use Enpii_Base\App\Support\App_Const;
 use Enpii_Base\Foundation\Http\Base_Controller;
 use Illuminate\Support\Facades\Artisan;
 
@@ -16,15 +17,7 @@ class Main_Controller extends Base_Controller {
 	 * @throws \Exception
 	 */
 	public function setup() {
-		enpii_base_wp_app_prepare_folders();
-
-		Artisan::call(
-			'wp-app:setup',
-			[]
-		);
-		$output = Artisan::output();
-		echo( nl2br( esc_html( $output ) ) );
-
-		return 'Setup';
+		do_action( App_Const::ACTION_WP_APP_SETUP_APP );
+		return 'Complete Setup';
 	}
 }
