+++ src/App/Http/Controllers/Main_Controller.php
@@ -7,28 +7,10 @@ namespace Enpii_Base\App\Http\Controllers;
 use Enpii_Base\App\Models\Post;
 use Enpii_Base\App\WP\Enpii_Base_WP_Plugin;
 use Enpii_Base\Foundation\Http\Base_Controller;
-use Illuminate\Support\Facades\Artisan;
 
 class Main_Controller extends Base_Controller {
 	public function index() {
 		return Enpii_Base_WP_Plugin::wp_app_instance()->view( 'main/index' );
-	}
-
-	public function queue_work() {
-		Artisan::call(
-			'queue:work',
-			[
-				'connection' => 'database',
-				'--queue' => 'high,default,low',
-				'--tries' => 3,
-				'--quiet' => true,
-				'--stop-when-empty' => true,
-				'--timeout' => 60,
-				'--memory' => 256,
-			]
-		);
-
-		return ' Queue Work ';
 	}
 
 	public function post() {
@@ -57,7 +39,7 @@ class Main_Controller extends Base_Controller {
 		$wp_query = new \WP_Query(
 			[
 				'p' => $post->ID,
-			] 
+			]
 		);
 		$wp_query->the_post();
 
@@ -70,7 +52,7 @@ class Main_Controller extends Base_Controller {
 		$wp_query = new \WP_Query(
 			[
 				'page_id' => 2,
-			] 
+			]
 		);
 		$wp_query->the_post();
