+++ src/App/Jobs/Show_Admin_Notice_And_Disable_Plugin_Job.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Enpii_Base\App\Jobs;
 
-use Enpii_Base\Foundation\Bus\Dispatchable_Trait;
 use Enpii_Base\Foundation\Shared\Base_Job;
+use Enpii_Base\Foundation\Support\Executable_Trait;
 use Enpii_Base\Foundation\WP\WP_Plugin_Interface;
 use Illuminate\Support\Facades\Session;
 
 class Show_Admin_Notice_And_Disable_Plugin_Job extends Base_Job {
-	use Dispatchable_Trait;
+	use Executable_Trait;
 
 	/**
 	 * @var \Enpii_Base\Foundation\WP\WP_Plugin
