+++ src/App/Jobs/Show_Admin_Notice_From_Flash_Messages_Job.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Enpii_Base\App\Jobs;
 
-use Enpii_Base\Foundation\Bus\Dispatchable_Trait;
 use Enpii_Base\Foundation\Shared\Base_Job;
+use Enpii_Base\Foundation\Support\Executable_Trait;
 use Illuminate\Support\Facades\Session;
 
 class Show_Admin_Notice_From_Flash_Messages_Job extends Base_Job {
-	use Dispatchable_Trait;
+	use Executable_Trait;
 
 	/**
 	 * We want to show Admin notice for 4 types of messages: 'error', 'success', 'warning', 'info'
