+++ src/App/Jobs/Write_Queue_Work_Script_Job.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Enpii_Base\App\Jobs;
 
-use Enpii_Base\Foundation\Bus\Dispatchable_Trait;
 use Enpii_Base\Foundation\Shared\Base_Job;
+use Enpii_Base\Foundation\Support\Executable_Trait;
 use Illuminate\Contracts\Container\BindingResolutionException;
 use PHPUnit\Framework\ExpectationFailedException;
 use Exception;
 
 class Write_Queue_Work_Script_Job extends Base_Job {
-	use Dispatchable_Trait;
+	use Executable_Trait;
 
 	/**
 	 * Write a js to have periodly ajax request to handle the queue
@@ -23,7 +23,7 @@ class Write_Queue_Work_Script_Job extends Base_Job {
 	 */
 	public function handle(): void {
 		// We want to add the trailing slash to avoid the redirect in WP webserver rule
-		$queue_work_url = esc_js( wp_app_route_wp_url( 'wp-app-queue-work' ) . '/' );
+		$queue_work_url = esc_js( wp_app_route_wp_url( 'wp-api-queue-work' ) . '/' );
 
 		// We want to have an interval that works every 2 mins (120 000 miliseconds)
 		//  To perform the queue execution because we set timeout
@@ -43,7 +43,7 @@ class Write_Queue_Work_Script_Job extends Base_Job {
 			}
 			var ajax_request_to_queue_work_interval = window.setInterval(function(){
 				ajax_request_to_queue_work();
-			}, 120000);
+			}, 300000);
 			window.setTimeout(function() {
 				ajax_request_to_queue_work();
 			}, 12000);
