+++ src/App/WP/Enpii_Base_WP_Plugin.php
@@ -6,11 +6,11 @@ namespace Enpii_Base\App\WP;
 
 use Enpii_Base\App\Jobs\Conclude_WP_App_Request_Job;
 use Enpii_Base\App\Jobs\Init_WP_App_Bootstrap_Job;
+use Enpii_Base\App\Jobs\Perform_Setup_WP_App_Job;
 use Enpii_Base\App\Jobs\Process_WP_Api_Request_Job;
 use Enpii_Base\App\Jobs\Process_WP_App_Request_Job;
 use Enpii_Base\App\Jobs\Register_Base_WP_Api_Routes_Job;
 use Enpii_Base\App\Jobs\Register_Base_WP_App_Routes_Job;
-use Enpii_Base\App\Jobs\Register_Main_Service_Providers_Job;
 use Enpii_Base\App\Jobs\Show_Admin_Notice_From_Flash_Messages_Job;
 use Enpii_Base\App\Jobs\Write_Queue_Work_Script_Job;
 use Enpii_Base\App\Jobs\Write_Setup_Client_Script_Job;
@@ -70,6 +70,18 @@ final class Enpii_Base_WP_Plugin extends WP_Plugin {
 		parent::boot();
 	}
 
+	public function get_name(): string {
+		return 'Enpii Base';
+	}
+
+	public function get_version(): string {
+		return ENPII_BASE_PLUGIN_VERSION;
+	}
+
+	public function get_text_domain(): string {
+		return 'enpii';
+	}
+
 	public function manipulate_hooks(): void {
 		/** WP CLI */
 		add_action( 'cli_init', [ $this, 'register_wp_cli_commands' ] );
@@ -80,6 +92,8 @@ final class Enpii_Base_WP_Plugin extends WP_Plugin {
 
 		add_action( App_Const::ACTION_WP_APP_REGISTER_ROUTES, [ $this, 'register_base_wp_app_routes' ] );
 		add_action( App_Const::ACTION_WP_API_REGISTER_ROUTES, [ $this, 'register_base_wp_api_routes' ] );
+		add_action( App_Const::ACTION_WP_APP_QUEUE_WORK, [ $this, 'queue_work' ] );
+		add_action( App_Const::ACTION_WP_APP_SETUP_APP, [ $this, 'setup_app' ] );
 
 		add_filter( App_Const::FILTER_WP_APP_MAIN_SERVICE_PROVIDERS, [ $this, 'register_telescope_tinker' ] );
 
@@ -95,16 +109,8 @@ final class Enpii_Base_WP_Plugin extends WP_Plugin {
 		add_action( 'admin_head', [ $this, 'handle_admin_head' ] );
 	}
 
-	public function get_name(): string {
-		return 'Enpii Base';
-	}
-
-	public function get_version(): string {
-		return ENPII_BASE_PLUGIN_VERSION;
-	}
-
-	public function get_text_domain(): string {
-		return 'enpii';
+	public function setup_app(): void {
+		Perform_Setup_WP_App_Job::dispatchSync();
 	}
 
 	public function bootstrap_wp_app(): void {
@@ -112,11 +118,11 @@ final class Enpii_Base_WP_Plugin extends WP_Plugin {
 	}
 
 	public function write_setup_wp_app_client_script(): void {
-		Write_Setup_Client_Script_Job::dispatchSync();
+		Write_Setup_Client_Script_Job::execute_now();
 	}
 
 	public function write_queue_work_client_script(): void {
-		Write_Queue_Work_Script_Job::dispatchSync();
+		Write_Queue_Work_Script_Job::execute_now();
 	}
 
 	public function register_base_wp_app_routes(): void {
@@ -257,7 +263,11 @@ final class Enpii_Base_WP_Plugin extends WP_Plugin {
 	 * @throws BindingResolutionException
 	 */
 	public function handle_admin_head() {
-		Show_Admin_Notice_From_Flash_Messages_Job::dispatchSync();
+		Show_Admin_Notice_From_Flash_Messages_Job::execute_now();
+	}
+
+	public function queue_work() {
+		Show_Admin_Notice_From_Flash_Messages_Job::execute_now();
 	}
 
 	/**
